+++ ShipManagementAPI/shipmanagement.viewmodels/Ship.cs
@@ -16,10 +16,10 @@ namespace shipmanagement.viewmodels
 
         [Required(ErrorMessage = "Length is required")]
         [Range(1,int.MaxValue,ErrorMessage = "Length should be between 1 - 2,147,483,647")]
-        public int Length { get; set; }
+        public decimal Length { get; set; }
         [Range(1, int.MaxValue, ErrorMessage = "Width should be between 1 - 2,147,483,647")]
         [Required(ErrorMessage = "Width is required")]
-        public int Width { get; set; }
+        public decimal Width { get; set; }
 
         [Required(ErrorMessage = "Code is required")]
         [MaxLength(250, ErrorMessage = "Code cannot be greater than 15 characters")]
